% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adj_to_edgelist}
\alias{adj_to_edgelist}
\title{Converts an adjacency matrix to edgelist and edgeweight using the
\code{igraph} package.}
\usage{
adj_to_edgelist(adj, directed = TRUE, weighted = TRUE)
}
\arguments{
\item{adj}{Adjacency matrix of a network.}

\item{directed}{Logical, whether the network is directed. This value is
passed to \code{igraph::graph_from_adjacency_matrix()}.}

\item{weighted}{Logical, whether the network is weighted.}
}
\value{
A list of edgelist, edgeweight and directed.
}
\description{
Converts an adjacency matrix to edgelist and edgeweight using the
\code{igraph} package.
}
\keyword{internal}
