% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edge_to_adj}
\alias{edge_to_adj}
\title{Convert edgelist and edgeweight to adjacency matrix.}
\usage{
edge_to_adj(edgelist, edgeweight = NULL, directed = TRUE)
}
\arguments{
\item{edgelist}{A two column matrix represents edges.}

\item{edgeweight}{A vector represents the weight of edges. If \code{NULL},
all the edges are considered have weight 1.}

\item{directed}{Logical, whether the network is directed.}
}
\value{
An adjacency matrix.
}
\description{
Convert edgelist and edgeweight to adjacency matrix.
}
