% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edgelist_to_adj}
\alias{edgelist_to_adj}
\title{Convert edgelist and edgeweight to adjacency matrix.}
\usage{
edgelist_to_adj(edgelist, edgeweight, directed = TRUE)
}
\arguments{
\item{edgelist}{A two column matrix representing edges.}

\item{edgeweight}{A vector representing the weight of edges. If \code{NULL},
all edges are considered to have a weight of 1.}

\item{directed}{Logical, whether the network is directed.}
}
\value{
Returns an adjacency matrix.
}
\description{
Convert edgelist and edgeweight to adjacency matrix.
}
\keyword{internal}
