% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fill_weight_cpp}
\alias{fill_weight_cpp}
\title{Fill edgeweight into the adjacency matrix.
Defined for function \code{edge_to_adj}.}
\usage{
fill_weight_cpp(adj, edgelist, edgeweight)
}
\arguments{
\item{adj}{An adjacency matrix.}

\item{edgelist}{A two column matrix represents the edgelist.}

\item{edgeweight}{A vector represents the weight of edges.}
}
\value{
Adjacency matrix with edge weight.
}
\description{
Fill edgeweight into the adjacency matrix.
Defined for function \code{edge_to_adj}.
}
\keyword{internal}
