% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_node_undirected_cpp}
\alias{find_node_undirected_cpp}
\title{Fill missing values in node sequence. Defined for \code{wdnet::rpanet}.}
\usage{
find_node_undirected_cpp(node1, node2, start_edge, end_edge)
}
\arguments{
\item{node1}{Nodes in the first column of edgelist, i.e., \code{edgelist[, 1]}.}

\item{node2}{Nodes in the second column of edgelist, i.e., \code{edgelist[, 2]}.}

\item{start_edge}{Index of sampled edges, corresponds to the missing nodes in node1 and node2.}

\item{end_edge}{Index of sampled edges, corresponds to the missing nodes in node1 and node2.}
}
\value{
Node sequence.
}
\description{
Fill missing values in node sequence. Defined for \code{wdnet::rpanet}.
}
