% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findNode_cpp}
\alias{findNode_cpp}
\title{Fill missing nodes in the node sequence. Defined for \code{wdnet::rpanet}.}
\usage{
findNode_cpp(nodes, edges)
}
\arguments{
\item{nodes}{Source/target nodes, missing nodes are denoted as 0.}

\item{edges}{Sampled edges according to preferential attachment.}
}
\value{
Source/target nodes.
}
\description{
Fill missing nodes in the node sequence. Defined for \code{wdnet::rpanet}.
}
