% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_dist.R
\name{get_constr}
\alias{get_constr}
\title{Get the constraints for the optimization problem. This function is defined
for \code{get_eta_directed()}.}
\usage{
get_constr(constrs, target.assortcoef, rho)
}
\arguments{
\item{constrs}{A list of constraints.}

\item{target.assortcoef}{A list of target assortativity levels.}

\item{rho}{A list of variable objects.}
}
\value{
A list of constraints.
}
\description{
Get the constraints for the optimization problem. This function is defined
for \code{get_eta_directed()}.
}
\keyword{internal}
