% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_dist.R
\name{get_dist}
\alias{get_dist}
\title{Get the node-level joint distributions and some empirical distributions with
given edgelist.}
\usage{
get_dist(edgelist = NA, directed = TRUE, joint_dist = FALSE)
}
\arguments{
\item{edgelist}{A two column matrix represents the directed edges of a
network.}

\item{directed}{Logical, whether the network is directed.}

\item{joint_dist}{Logical, whether to return edge-level distributions.}
}
\value{
A list of distributions and degree vectors.
}
\description{
Get the node-level joint distributions and some empirical distributions with
given edgelist.
}
