% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_dist.R
\name{get_values}
\alias{get_values}
\title{Get the value of an object from the optimization problem. This function is
defined for \code{get_eta_directed()}.}
\usage{
get_values(object, result, mydist)
}
\arguments{
\item{object}{An object from the optimization problem.}

\item{result}{A list returned from \code{CVXR::solve()}.}

\item{mydist}{A list returned from \code{get_dist()}.}
}
\value{
Returns the value of the object.
}
\description{
Get the value of an object from the optimization problem. This function is
defined for \code{get_eta_directed()}.
}
\keyword{internal}
