% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_wdnet.R
\name{igraph_to_wdnet}
\alias{igraph_to_wdnet}
\title{Converts an \code{igraph} object to a \code{wdnet} object}
\usage{
igraph_to_wdnet(g)
}
\arguments{
\item{g}{An \code{igraph} object.}
}
\value{
A \code{wdnet} object.
}
\description{
Converts an \code{igraph} object to a \code{wdnet} object
}
\examples{
g <- igraph::sample_pa(50)
netwk <- igraph_to_wdnet(g)

}
