% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_wdnet.R
\name{is_wdnet}
\alias{is_wdnet}
\title{Checks if the input is a \code{wdnet} object}
\usage{
is_wdnet(netwk)
}
\arguments{
\item{netwk}{A \code{wdnet} object.}
}
\value{
Logical, \code{TRUE} if argument netwk is a \code{wdnet} object.
}
\description{
Checks if the input is a \code{wdnet} object
}
\examples{
netwk <- rpanet(nstep = 1e3)
is_wdnet(netwk)

}
