% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{node_strength_cpp}
\alias{node_strength_cpp}
\title{Aggregate edgeweight into nodes' strength.}
\usage{
node_strength_cpp(snode, tnode, weight, nnode, weighted = TRUE)
}
\arguments{
\item{snode}{Source nodes.}

\item{tnode}{Target nodes.}

\item{weight}{Edgeweight.}

\item{nnode}{Number of nodes.}

\item{weighted}{Logical, true if the edges are weighted, 
false if not.}
}
\value{
Out-strength and in-strength.
}
\description{
Aggregate edgeweight into nodes' strength.
}
\keyword{internal}
