% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_wdnet.R
\name{plot.wdnet}
\alias{plot.wdnet}
\title{Plots the input network}
\usage{
\method{plot}{wdnet}(x, ...)
}
\arguments{
\item{x}{A \code{wdnet} object.}

\item{...}{Additional parameters passed to \code{igraph::plot.igraph()}.}
}
\value{
Returns \code{NULL}, invisibly.
}
\description{
Plots the input network via \code{igraph::plot.igraph()}.
}
