% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_rpacontrol.R
\name{print.rpacontrol}
\alias{print.rpacontrol}
\alias{summary.rpacontrol}
\title{Prints \code{rpacontrol} objects}
\usage{
\method{print}{rpacontrol}(x, ...)

\method{summary}{rpacontrol}(object, ...)
}
\arguments{
\item{x}{An object of class \code{rpacontrol}.}

\item{...}{Additional arguments.}

\item{object}{An object of class \code{rpacontrol}.}
}
\value{
Returns the controls invisibly.
}
\description{
These functions print \code{rpacontrol} objects in the terminal.
\code{print.rpacontrol()} shows only the current controls, whereas
\code{summary.rpacontrol()} includes both specified controls and the
unspecified controls that use default values.
}
\examples{

control <- rpa_control_scenario()
print(control)

}
