% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_rpacontrol.R
\name{print_control_edgeweight}
\alias{print_control_edgeweight}
\title{Prints \code{rpa_control_edgeweight()} in terminal}
\usage{
print_control_edgeweight(control)
}
\arguments{
\item{control}{A list of control parameters for
\code{rpa_control_edgeweight()}.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Prints \code{rpa_control_edgeweight()} in terminal
}
\keyword{internal}
