% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_rpacontrol.R
\name{print_control_preference}
\alias{print_control_preference}
\title{Prints preference functions in terminal}
\usage{
print_control_preference(control, directed = NULL)
}
\arguments{
\item{control}{An object of class \code{rpacontrol}.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Prints preference functions in terminal
}
\keyword{internal}
