% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_rpacontrol.R
\name{print_control_reciprocal}
\alias{print_control_reciprocal}
\title{Prints \code{rpa_control_reciprocal()} in terminal}
\usage{
print_control_reciprocal(control)
}
\arguments{
\item{control}{A list of control parameters for
\code{rpa_control_reciprocal()}.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Prints \code{rpa_control_reciprocal()} in terminal
}
\keyword{internal}
