% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_rpacontrol.R
\name{print_control_scenario}
\alias{print_control_scenario}
\title{Prints \code{rpa_control_scenario()} in terminal}
\usage{
print_control_scenario(control)
}
\arguments{
\item{control}{A list of control parameters for
\code{rpa_control_scenario()}.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Prints \code{rpa_control_scenario()} in terminal
}
\keyword{internal}
