% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpa_control.R
\name{rpa_control_default}
\alias{rpa_control_default}
\title{Default controls for \code{rpanet}}
\usage{
rpa_control_default()
}
\value{
Returns a list of default controls.
}
\description{
Default controls for \code{rpanet}
}
\keyword{internal}
