% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpa_control.R
\name{rpa_control_list}
\alias{rpa_control_list}
\title{Names and descriptions of controls for \code{rpanet()}}
\usage{
rpa_control_list()
}
\value{
Returns a list of control names and descriptions.
}
\description{
Names and descriptions of controls for \code{rpanet()}
}
\keyword{internal}
