% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpa_control.R
\name{rpacontrol}
\alias{rpacontrol}
\title{rpacontrol: Controls the Preferential Attachment (PA) Network Generation
Process}
\description{
The \code{rpacontrol} object is designed to control the Preferential
Attachment (PA) network generation process within the \code{rpanet()}
function. It can have the following components:
\itemize{
\item \code{scenario}: controls the edge scenarios
at each step. For more information, please refer to
\code{rpa_control_scenario()}.
\item \code{edgeweight}: controls the weight of
the edges; see \code{rpa_control_edgeweight()} for details.
\item \code{newedge}: controls the creation of
new edges at each step; see \code{rpa_control_newedge()}
for details.
\item \code{preference}: sets preference functions; see
\code{rpa_control_preference()} for details.
\item \code{reciprocal}: controls the creation of reciprocal
edges; see \code{rpa_control_reciprocal()} for details.
}
}
