% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rpanet_binary_directed}
\alias{rpanet_binary_directed}
\title{Preferential attachment algorithm.}
\usage{
rpanet_binary_directed(
  nstep,
  m,
  new_node_id,
  new_edge_id,
  source_node,
  target_node,
  outs,
  ins,
  edgeweight,
  scenario,
  sample_recip,
  node_group,
  source_pref,
  target_pref,
  control
)
}
\arguments{
\item{nstep}{Number of steps.}

\item{m}{Number of new edges in each step.}

\item{new_node_id}{New node ID.}

\item{new_edge_id}{New edge ID.}

\item{source_node}{Sequence of source nodes.}

\item{target_node}{Sequence of target nodes.}

\item{outs}{Sequence of out-strength.}

\item{ins}{Sequence of in-strength.}

\item{edgeweight}{Weight of existing and new edges.}

\item{scenario}{Scenario of existing and new edges.}

\item{sample_recip}{Logical, whether reciprocal edges will be added.}

\item{node_group}{Sequence of node group.}

\item{source_pref}{Sequence of node source preference.}

\item{target_pref}{Sequence of node target preference.}

\item{control}{List of controlling arguments.}
}
\value{
Sampled network.
}
\description{
Preferential attachment algorithm.
}
