% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rpanet_linear_undirected_cpp}
\alias{rpanet_linear_undirected_cpp}
\title{Preferential attachment algorithm.}
\usage{
rpanet_linear_undirected_cpp(
  nstep,
  m,
  new_node_id,
  new_edge_id,
  node_vec1,
  node_vec2,
  s,
  edgeweight,
  scenario,
  pref_vec,
  control
)
}
\arguments{
\item{nstep}{Number of steps.}

\item{m}{Number of new edges in each step.}

\item{new_node_id}{New node ID.}

\item{new_edge_id}{New edge ID.}

\item{node_vec1}{Sequence of nodes in the first column of edgelist.}

\item{node_vec2}{Sequence of nodes in the second column of edgelist.}

\item{s}{Sequence of node strength.}

\item{edgeweight}{Weight of existing and new edges.}

\item{scenario}{Scenario of existing and new edges.}

\item{control}{List of controlling arguments.}

\item{pref}{Sequence of node preference.}
}
\value{
Sampled network.
}
\description{
Preferential attachment algorithm.
}
\keyword{internal}
