% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rpanet_nodelist_cpp}
\alias{rpanet_nodelist_cpp}
\title{Preferential attachment algorithm for simple situations, 
i.e., edge weight equals to 1, number of new edges per step is 1.}
\usage{
rpanet_nodelist_cpp(
  snode,
  tnode,
  scenario,
  nnode,
  nedge,
  delta_out,
  delta_in,
  directed
)
}
\arguments{
\item{snode}{Source nodes.}

\item{tnode}{Target nodes.}

\item{scenario}{Sequence of alpha, beta, gamma, xi, rho scenarios.}

\item{nnode}{Number of nodes in seed network.}

\item{nedge}{Number of edges in seed network.}

\item{delta_out}{Tuning parameter.}

\item{delta_in}{Tuning parameter.}

\item{directed}{Whether the network is directed.}
}
\value{
Number of nodes, sequences of source and target nodes.
}
\description{
Preferential attachment algorithm for simple situations, 
i.e., edge weight equals to 1, number of new edges per step is 1.
}
