% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpanet_wan.R
\name{rpanet_wan}
\alias{rpanet_wan}
\title{Simulating a Preferential Attachment Network}
\usage{
rpanet_wan(alpha, beta, gamma, xi, delta_in, delta_out, nedge)
}
\arguments{
\item{alpha}{Scalar probability of adding an edge from the new node
to an existing node}

\item{beta}{Scalar probability of adding an edge between two
existing nodes.}

\item{gamma}{Scalar probability of adding an edge from an existing
node to a new node.}

\item{xi}{Scalar probability of ...}

\item{delta_in}{Growth rate parameter for nodes' instrength}

\item{delta_out}{Growth rate parameter for nodes' outstrength}

\item{nedge}{The number of edges to be generated}
}
\value{
A list with the following components: in_degree,
    out_degree, edge_start, edge_end, evolution
}
\description{
Simulating a Preferential Attachment Network
}
