% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleNode_cpp}
\alias{sampleNode_cpp}
\title{Uniformly draw a node from existing nodes for each time step.
Defined for \code{wdnet::rpanet}.}
\usage{
sampleNode_cpp(total_node)
}
\arguments{
\item{total_node}{Number of existing nodes at each time step.}
}
\value{
Sampled nodes.
}
\description{
Uniformly draw a node from existing nodes for each time step.
Defined for \code{wdnet::rpanet}.
}
