% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdnet-package.R
\docType{package}
\name{wdnet}
\alias{wdnet}
\title{wdnet: Weighted and Directed Networks}
\description{
This package provides functions to conduct network analysis
\itemize{
\item Assortativity, centrality, clustering coefficient
for weighted and directed networks
\item Rewire an unweighted network with given assortativity coefficient(s)
\item Preferential attachment (PA) network generation
}
}
\details{
The development version of this package is available on Gitlab
(\url{https://gitlab.com/wdnetwork/wdnet}).
}
