% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_wdnet.R
\name{wdnet_to_igraph}
\alias{wdnet_to_igraph}
\title{Converts a \code{wdnet} object to an \code{igraph} object}
\usage{
wdnet_to_igraph(netwk)
}
\arguments{
\item{netwk}{A \code{wdnet} object.}
}
\value{
An \code{igraph} object.
}
\description{
Converts a \code{wdnet} object to an \code{igraph} object
}
\examples{
netwk <- rpanet(nstep = 1e3)
g <- wdnet_to_igraph(netwk)

}
