% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_pref.R
\name{compile_pref_func}
\alias{compile_pref_func}
\title{Compile preference functions via \code{Rcpp}.}
\usage{
compile_pref_func(preference)
}
\arguments{
\item{preference}{A list for defining the preference functions.}
}
\value{
Preference functions and external pointers.
}
\description{
Compile preference functions via \code{Rcpp}.
}
\keyword{internal}
