% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_pref.R
\name{compile_pref_func}
\alias{compile_pref_func}
\title{Compile preference functions via \code{RcppXPtrUtils}.}
\usage{
compile_pref_func(preference, directed)
}
\arguments{
\item{preference}{A list for defining the preference functions.}

\item{directed}{Logical, whether to compile the preference functions for
directed networks. If missing, the preference functions
for both directed and undirected networks will be compiled.}
}
\value{
Returns the input list and their corresponding external pointers.
}
\description{
Compile preference functions via \code{RcppXPtrUtils}.
}
\keyword{internal}
