% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_rpacontrol.R
\name{is_rpacontrol}
\alias{is_rpacontrol}
\title{Checks whether the input is a \code{rpacontrol} object}
\usage{
is_rpacontrol(control)
}
\arguments{
\item{control}{A \code{rpacontrol} object.}
}
\value{
Logical, \code{TRUE} if the input is a \code{rpacontrol} object.
}
\description{
Checks whether the input is a \code{rpacontrol} object
}
\keyword{internal}
