% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_rpacontrol.R
\name{print_control_details}
\alias{print_control_details}
\title{Prints \code{rpacontrol} in terminal}
\usage{
print_control_details(x, control_name)
}
\arguments{
\item{x}{An object of class \code{rpacontrol}.}

\item{control_name}{A string, the name of the control component.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Prints \code{rpacontrol} in terminal
}
\keyword{internal}
